+++ flutter_local_notifications/CHANGELOG.md
@@ -1,9 +1,32 @@
-## [vNext]
+## [19.0.0-dev.1]
 
-* **Breaking change** Bumped minimum Flutter SDK requirement to 3.13
+* **Breaking change** bumped minimum Flutter SDK requirement to 3.19.0 and Dart SDK requirement to 3.3.0
+* **Breaking change** [iOS] removed `uiLocalNotificationDateInterpretation` parameter from `zonedSchedule()` method. This was done as it actually no relevant as of the 18.0.0 that dropped support for iOS versions older than 10, which in turn meant that the deprecated `UILocalNotification` APIs from Apple were no longer used. The corresponding `UILocalNotificationDateInterpretation` enum has already been removed as well
+* [Windows] Added support for Windows. Thanks to the PR from [Levi Lesches](https://github.com/Levi-Lesches/)
+* Bumped `timezone` dependency so that minimum version is now 0.10.0
+
+## [18.0.1]
+
+* Added upper bound constraints for [timezone](https://pub.dev/packages/timezone) dependency. This is to allow the plugin and apps pull version 0.10.x if constraints are satisified. Future releases may bump the minimum `timezone` dependency but this approach is being used at the moment as version 0.10.0 of the time `timezone` package introduces a dependency on `http` package. Directly bumping the `flutter_local_notifications` dependency on `timezone` could have introduced dependency conflicts
+
+## [18.0.0]
+
+* **Breaking changes** Bumped minimum Flutter SDK requirement to 3.13. Consequently the minimum OS requirements for each platform has been updated as well
+  * [Android] minimum Android version is now 4.4 (API level 19)
+  * [iOS] minimum iOS version is now 11
+  * [macOS] minimum macOS version is now 10.14
 * [Android] **Breaking change**  removed the deprecated `androidAllowWhileIdle` parameter from `zonedSchedule()` and `periodicallyShow()` methods. `androidScheduleMode` is now a required parameter
+* **Breaking change** plugin has been fixed with regards to how it registers the platform-specific implementations of the plugin. Thanks to the PR from [Kate](https://github.com/provokateurin). Any written tests done on the [FlutterLocalNotificationsPlugin] will need to manually call the `registerWith()` method that has been added to each implementation. This affects the following platforms/classes where the `registerWith()` method was added
+  * [Android] `AndroidFlutterLocalNotificationsPlugin`
+  * [iOS] `IOSFlutterLocalNotificationsPlugin`
+  * [macOS] `MacOSFlutterLocalNotificationsPlugin`
+* [iOS] **Breaking change** Removed `onDidReceiveLocalNotification` callback as this was only relevant on iOS versions older than 10
 * Fixed example app to have the appropriate permissions for foreground services
 * Updated readme when it comes to setting up the `AndroidManifest.xml` file to include details of what's needed for foreground services. Note these details were already available in the API docs
+
+## [17.2.4]
+
+* [macOS] added privacy manifest file
 
 ## [17.2.3]