+++ flutter_local_notifications_platform_interface/CHANGELOG.md
@@ -1,4 +1,8 @@
-## [vNext]
+## [8.1.0-dev.1]
+
+* Added `data` property to `NotificationResponse` class to support notification actions on Windows. Thanks to the PR from [Levi Lesches](https://github.com/Levi-Lesches)
+
+## [8.0.0]
 
 * **Breaking change** Bumped minimum Flutter SDK requirement to 3.13